+++ src/main/java/com/unitechApi/store/indent/Service/IndentService.java
@@ -31,7 +31,7 @@ public class IndentService {
     public Indent saveData(Indent indent ) {
         StoreItemModel itemModel=storeItemRepository.findById(indent.getStoreItem().getItemId())
                 .orElseThrow(()-> new ItemNotFound("item not Found"));
-        indent.setTotal(indent.getEstimatedPrice() * indent.getQuantity());
+        indent.setTotal((long) (indent.getEstimatedPrice() * indent.getQuantity()));
         indent.setIncludingTax(((indent.getTotal() * itemModel.getPaytax())/100)+indent.getTotal());
    //     indent.set((long) (indent.getTotal() + indent.getIncludingTax()));
         return indentRepository.save(indent);
